+++ pandapower/control/controller/trafo/TapDependentImpedance.py
@@ -26,14 +26,14 @@ class TapDependentImpedance(CharacteristicControl):
         **drop_same_existing_ctrl** (bool, False) - Indicates if already existing controllers of the same type and with the same matching parameters (e.g. at same element) should be dropped
     """
 
-    def __init__(self, net, transformer_index, characteristic, trafotable="trafo", output_variable="vk_percent", tol=1e-3, restore=True,
-                 in_service=True, order=0, level=0, drop_same_existing_ctrl=False, matching_params=None, **kwargs):
+    def __init__(self, net, transformer_index, characteristic_index, trafotable="trafo", output_variable="vk_percent", tol=1e-3,
+                 restore=True, in_service=True, order=0, level=0, drop_same_existing_ctrl=False, matching_params=None, **kwargs):
         if matching_params is None:
             matching_params = {"transformer_index": transformer_index, 'output_variable': output_variable}
         super().__init__(net, output_element=trafotable, output_variable=output_variable, output_element_index=transformer_index,
                          input_element=trafotable, input_variable="tap_pos", input_element_index=transformer_index,
-                         characteristic=characteristic, tol=tol, in_service=in_service, order=order,
-                         level=level, drop_same_existing_ctrl=drop_same_existing_ctrl, matching_params=matching_params, **kwargs)
+                         characteristic_index=characteristic_index, tol=tol, in_service=in_service, order=order, level=level,
+                         drop_same_existing_ctrl=drop_same_existing_ctrl, matching_params=matching_params, **kwargs)
         self.restore=restore
         self.initial_values = net[trafotable].loc[transformer_index, output_variable].copy()