+++ pandapower/create.py
@@ -2120,7 +2120,9 @@ def create_lines_from_parameters(net, from_buses, to_buses, length_km, r_ohm_per
 
 
 def create_transformer(net, hv_bus, lv_bus, std_type, name=None, tap_pos=nan, in_service=True,
-                       index=None, max_loading_percent=nan, parallel=1, df=1.):
+                       index=None, max_loading_percent=nan, parallel=1, df=1.,
+                       tap_dependent_impedance=None, vk_percent_characteristic=None,
+                       vkr_percent_characteristic=None):
     """
     Creates a two-winding transformer in table net["trafo"].
     The trafo parameters are defined through the standard type library.
@@ -2167,6 +2169,20 @@ def create_transformer(net, hv_bus, lv_bus, std_type, name=None, tap_pos=nan, in
 
         **df** (float) - derating factor: maximal current of transformer in relation to nominal \
             current of transformer (from 0 to 1)
+
+        **tap_dependent_impedance** (boolean) - True if transformer impedance must be adjusted dependent \
+            on the tap position of the trabnsformer. Requires the additional columns \
+            "vk_percent_characteristic" and "vkr_percent_characteristic" that reference the index of the \
+            characteristic from the table net.characteristic. A convenience function \
+            pandapower.control.create_trafo_characteristics can be used to create the SplineCharacteristic \
+            objects, add the relevant columns and set up the references to the characteristics. \
+            The function pandapower.control.trafo_characteristics_diagnostic can be used for sanity checks.
+
+        **vk_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vkr_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
 
     OUTPUT:
         **index** (int) - The unique ID of the created transformer
@@ -2220,6 +2236,13 @@ def create_transformer(net, hv_bus, lv_bus, std_type, name=None, tap_pos=nan, in
 
     _create_column_and_set_value(net, index, max_loading_percent, "max_loading_percent", "trafo")
 
+    if tap_dependent_impedance is not None:
+        _create_column_and_set_value(net, index, tap_dependent_impedance, "tap_dependent_impedance", "trafo", bool_, False, True)
+    if vk_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vk_percent_characteristic, "vk_percent_characteristic", "trafo")
+    if vkr_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vkr_percent_characteristic, "vkr_percent_characteristic", "trafo")
+
     # tap_phase_shifter default False
     net.trafo.tap_phase_shifter.fillna(False, inplace=True)
 
@@ -2237,7 +2260,9 @@ def create_transformer_from_parameters(net, hv_bus, lv_bus, sn_mva, vn_hv_kv, vn
                                        df=1., vk0_percent=nan, vkr0_percent=nan,
                                        mag0_percent=nan, mag0_rx=nan,
                                        si0_hv_partial=nan,
-                                       pt_percent=nan, oltc=False, **kwargs):
+                                       pt_percent=nan, oltc=False, tap_dependent_impedance=None,
+                                       vk_percent_characteristic=None,
+                                       vkr_percent_characteristic=None, **kwargs):
     """
     Creates a two-winding transformer in table net["trafo"].
     The trafo parameters are defined through the standard type library.
@@ -2317,6 +2342,20 @@ def create_transformer_from_parameters(net, hv_bus, lv_bus, sn_mva, vn_hv_kv, vn
         **df** (float) - derating factor: maximal current of transformer in relation to nominal \
             current of transformer (from 0 to 1)
 
+        **tap_dependent_impedance** (boolean) - True if transformer impedance must be adjusted dependent \
+            on the tap position of the trabnsformer. Requires the additional columns \
+            "vk_percent_characteristic" and "vkr_percent_characteristic" that reference the index of the \
+            characteristic from the table net.characteristic. A convenience function \
+            pandapower.control.create_trafo_characteristics can be used to create the SplineCharacteristic \
+            objects, add the relevant columns and set up the references to the characteristics. \
+            The function pandapower.control.trafo_characteristics_diagnostic can be used for sanity checks.
+
+        **vk_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vkr_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
         **pt_percent** (float, nan) - (short circuit only)
 
         **oltc** (bool, False) - (short circuit only)
@@ -2364,6 +2403,13 @@ def create_transformer_from_parameters(net, hv_bus, lv_bus, sn_mva, vn_hv_kv, vn
 
     _set_entries(net, "trafo", index, **v, **kwargs)
 
+    if tap_dependent_impedance is not None:
+        _create_column_and_set_value(net, index, tap_dependent_impedance, "tap_dependent_impedance", "trafo", bool_, False, True)
+    if vk_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vk_percent_characteristic, "vk_percent_characteristic", "trafo")
+    if vkr_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vkr_percent_characteristic, "vkr_percent_characteristic", "trafo")
+
     if not (isnan(vk0_percent) and isnan(vkr0_percent) and isnan(mag0_percent)
             and isnan(mag0_rx) and isnan(si0_hv_partial) and vector_group is None):
         _create_column_and_set_value(net, index, vk0_percent, "vk0_percent", "trafo")
@@ -2387,7 +2433,9 @@ def create_transformers_from_parameters(net, hv_buses, lv_buses, sn_mva, vn_hv_k
                                         vector_group=None, index=None, max_loading_percent=None,
                                         parallel=1, df=1., vk0_percent=None, vkr0_percent=None,
                                         mag0_percent=None, mag0_rx=None, si0_hv_partial=None,
-                                        pt_percent=nan, oltc=False, **kwargs):
+                                        pt_percent=nan, oltc=False, tap_dependent_impedance=None,
+                                        vk_percent_characteristic=None,
+                                        vkr_percent_characteristic=None, **kwargs):
     """
     Creates several two-winding transformers in table net["trafo"].
     The trafo parameters are defined through the standard type library.
@@ -2468,6 +2516,20 @@ def create_transformers_from_parameters(net, hv_buses, lv_buses, sn_mva, vn_hv_k
         **df** (float) - derating factor: maximal current of transformer in relation to nominal \
             current of transformer (from 0 to 1)
 
+        **tap_dependent_impedance** (boolean) - True if transformer impedance must be adjusted dependent \
+            on the tap position of the trabnsformer. Requires the additional columns \
+            "vk_percent_characteristic" and "vkr_percent_characteristic" that reference the index of the \
+            characteristic from the table net.characteristic. A convenience function \
+            pandapower.control.create_trafo_characteristics can be used to create the SplineCharacteristic \
+            objects, add the relevant columns and set up the references to the characteristics. \
+            The function pandapower.control.trafo_characteristics_diagnostic can be used for sanity checks.
+
+        **vk_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vkr_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
         **pt_percent** (float, nan) - (short circuit only)
 
         **oltc** (bool, False) - (short circuit only)
@@ -2498,6 +2560,13 @@ def create_transformers_from_parameters(net, hv_buses, lv_buses, sn_mva, vn_hv_k
                "tap_phase_shifter": tap_phase_shifter, "parallel": parallel, "df": df,
                "pt_percent": pt_percent, "oltc": oltc}
 
+    if tap_dependent_impedance is not None:
+        _add_series_to_entries(entries, index, "tap_dependent_impedance", tap_dependent_impedance, dtype=bool_, default_val=False)
+    if vk_percent_characteristic is not None:
+        _add_series_to_entries(entries, index, "vk_percent_characteristic",  vk_percent_characteristic)
+    if vkr_percent_characteristic is not None:
+        _add_series_to_entries(entries, index, "vkr_percent_characteristic", vkr_percent_characteristic)
+
     _add_series_to_entries(entries, index, "vk0_percent", vk0_percent)
     _add_series_to_entries(entries, index, "vkr0_percent", vkr0_percent)
     _add_series_to_entries(entries, index, "mag0_percent", mag0_percent)
@@ -2514,7 +2583,10 @@ def create_transformers_from_parameters(net, hv_buses, lv_buses, sn_mva, vn_hv_k
 
 def create_transformer3w(net, hv_bus, mv_bus, lv_bus, std_type, name=None, tap_pos=nan,
                          in_service=True, index=None, max_loading_percent=nan,
-                         tap_at_star_point=False):
+                         tap_at_star_point=False, tap_dependent_impedance=None,
+                         vk_hv_percent_characteristic=None, vkr_hv_percent_characteristic=None,
+                         vk_mv_percent_characteristic=None, vkr_mv_percent_characteristic=None,
+                         vk_lv_percent_characteristic=None, vkr_lv_percent_characteristic=None):
     """
     Creates a three-winding transformer in table net["trafo3w"].
     The trafo parameters are defined through the standard type library.
@@ -2549,6 +2621,32 @@ def create_transformer3w(net, hv_bus, mv_bus, lv_bus, std_type, name=None, tap_p
         **max_loading_percent (float)** - maximum current loading (only needed for OPF)
 
         **tap_at_star_point (bool)** - whether tap changer is modelled at star point or at the bus
+
+        **tap_dependent_impedance** (boolean) - True if transformer impedance must be adjusted dependent \
+            on the tap position of the trabnsformer. Requires the additional columns \
+            "vk_percent_characteristic" and "vkr_percent_characteristic" that reference the index of the \
+            characteristic from the table net.characteristic. A convenience function \
+            pandapower.control.create_trafo_characteristics can be used to create the SplineCharacteristic \
+            objects, add the relevant columns and set up the references to the characteristics. \
+            The function pandapower.control.trafo_characteristics_diagnostic can be used for sanity checks.
+
+        **vk_hv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vkr_hv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vk_mv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vkr_mv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vk_lv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vkr_lv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
 
     OUTPUT:
         **index** (int) - The unique ID of the created transformer
@@ -2613,6 +2711,21 @@ def create_transformer3w(net, hv_bus, mv_bus, lv_bus, std_type, name=None, tap_p
 
     _create_column_and_set_value(net, index, max_loading_percent, "max_loading_percent", "trafo3w")
 
+    if tap_dependent_impedance is not None:
+        _create_column_and_set_value(net, index, tap_dependent_impedance, "tap_dependent_impedance", "trafo", bool_, False, True)
+    if vk_hv_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vk_hv_percent_characteristic, "vk_hv_percent_characteristic", "trafo")
+    if vkr_hv_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vkr_hv_percent_characteristic, "vkr_hv_percent_characteristic", "trafo")
+    if vk_mv_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vk_mv_percent_characteristic, "vk_mv_percent_characteristic", "trafo")
+    if vkr_mv_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vkr_mv_percent_characteristic, "vkr_mv_percent_characteristic", "trafo")
+    if vk_lv_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vk_lv_percent_characteristic, "vk_lv_percent_characteristic", "trafo")
+    if vkr_lv_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vkr_lv_percent_characteristic, "vkr_lv_percent_characteristic", "trafo")
+
     return index
 
 
@@ -2627,7 +2740,10 @@ def create_transformer3w_from_parameters(net, hv_bus, mv_bus, lv_bus, vn_hv_kv, 
                                          max_loading_percent=nan, tap_at_star_point=False,
                                          vk0_hv_percent=nan, vk0_mv_percent=nan, vk0_lv_percent=nan,
                                          vkr0_hv_percent=nan, vkr0_mv_percent=nan, vkr0_lv_percent=nan,
-                                         vector_group=None):
+                                         vector_group=None, tap_dependent_impedance=None,
+                                         vk_hv_percent_characteristic=None, vkr_hv_percent_characteristic=None,
+                                         vk_mv_percent_characteristic=None, vkr_mv_percent_characteristic=None,
+                                         vk_lv_percent_characteristic=None, vkr_lv_percent_characteristic=None):
     """
     Adds a three-winding transformer in table net["trafo3w"].
     The model currently only supports one tap-changer per 3W Transformer.
@@ -2701,6 +2817,32 @@ def create_transformer3w_from_parameters(net, hv_bus, mv_bus, lv_bus, vn_hv_kv, 
 
         **max_loading_percent (float)** - maximum current loading (only needed for OPF)
 
+        **tap_dependent_impedance** (boolean) - True if transformer impedance must be adjusted dependent \
+            on the tap position of the trabnsformer. Requires the additional columns \
+            "vk_percent_characteristic" and "vkr_percent_characteristic" that reference the index of the \
+            characteristic from the table net.characteristic. A convenience function \
+            pandapower.control.create_trafo_characteristics can be used to create the SplineCharacteristic \
+            objects, add the relevant columns and set up the references to the characteristics. \
+            The function pandapower.control.trafo_characteristics_diagnostic can be used for sanity checks.
+
+        **vk_hv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vkr_hv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vk_mv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vkr_mv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vk_lv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vkr_lv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
         **vk0_hv_percent** (float) - zero sequence short circuit voltage from high to medium voltage
 
         **vk0_mv_percent** (float) - zero sequence short circuit voltage from medium to low voltage
@@ -2756,6 +2898,21 @@ def create_transformer3w_from_parameters(net, hv_bus, mv_bus, lv_bus, vn_hv_kv, 
 
     _create_column_and_set_value(net, index, max_loading_percent, "max_loading_percent", "trafo3w")
 
+    if tap_dependent_impedance is not None:
+        _create_column_and_set_value(net, index, tap_dependent_impedance, "tap_dependent_impedance", "trafo", bool_, False, True)
+    if vk_hv_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vk_hv_percent_characteristic, "vk_hv_percent_characteristic", "trafo")
+    if vkr_hv_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vkr_hv_percent_characteristic, "vkr_hv_percent_characteristic", "trafo")
+    if vk_mv_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vk_mv_percent_characteristic, "vk_mv_percent_characteristic", "trafo")
+    if vkr_mv_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vkr_mv_percent_characteristic, "vkr_mv_percent_characteristic", "trafo")
+    if vk_lv_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vk_lv_percent_characteristic, "vk_lv_percent_characteristic", "trafo")
+    if vkr_lv_percent_characteristic is not None:
+        _create_column_and_set_value(net, index, vkr_lv_percent_characteristic, "vkr_lv_percent_characteristic", "trafo")
+
     return index
 
 
@@ -2770,7 +2927,10 @@ def create_transformers3w_from_parameters(net, hv_buses, mv_buses, lv_buses, vn_
                                           tap_at_star_point=False,
                                           vk0_hv_percent=nan, vk0_mv_percent=nan, vk0_lv_percent=nan,
                                           vkr0_hv_percent=nan, vkr0_mv_percent=nan, vkr0_lv_percent=nan,
-                                          vector_group=None, **kwargs):
+                                          vector_group=None, tap_dependent_impedance=None,
+                                          vk_hv_percent_characteristic=None, vkr_hv_percent_characteristic=None,
+                                          vk_mv_percent_characteristic=None, vkr_mv_percent_characteristic=None,
+                                          vk_lv_percent_characteristic=None, vkr_lv_percent_characteristic=None, **kwargs):
     """
     Adds a three-winding transformer in table net["trafo3w"].
 
@@ -2849,6 +3009,32 @@ def create_transformers3w_from_parameters(net, hv_buses, mv_buses, lv_buses, vn_
 
         **max_loading_percent (float)** - maximum current loading (only needed for OPF)
 
+        **tap_dependent_impedance** (boolean) - True if transformer impedance must be adjusted dependent \
+            on the tap position of the trabnsformer. Requires the additional columns \
+            "vk_percent_characteristic" and "vkr_percent_characteristic" that reference the index of the \
+            characteristic from the table net.characteristic. A convenience function \
+            pandapower.control.create_trafo_characteristics can be used to create the SplineCharacteristic \
+            objects, add the relevant columns and set up the references to the characteristics. \
+            The function pandapower.control.trafo_characteristics_diagnostic can be used for sanity checks.
+
+        **vk_hv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vkr_hv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vk_mv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vkr_mv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vk_lv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
+        **vkr_lv_percent_characteristic** (int) - index of the characteristic from net.characteristic for \
+            the adjustment of the parameter "vk_percent" for the calculation of tap dependent impedance.
+
         **vk0_hv_percent** (float) - zero sequence short circuit voltage from high to medium voltage
 
         **vk0_mv_percent** (float) - zero sequence short circuit voltage from medium to low voltage
@@ -2907,6 +3093,21 @@ def create_transformers3w_from_parameters(net, hv_buses, mv_buses, lv_buses, vn_
                "vector_group": vector_group}
 
     _add_series_to_entries(entries, index, "max_loading_percent", max_loading_percent)
+
+    if tap_dependent_impedance is not None:
+        _add_series_to_entries(entries, index, "tap_dependent_impedance", tap_dependent_impedance, dtype=bool_, default_val=False)
+    if vk_hv_percent_characteristic is not None:
+        _add_series_to_entries(entries, index, "vk_hv_percent_characteristic", vk_hv_percent_characteristic)
+    if vkr_hv_percent_characteristic is not None:
+        _add_series_to_entries(entries, index, "vkr_hv_percent_characteristic", vkr_hv_percent_characteristic)
+    if vk_mv_percent_characteristic is not None:
+        _add_series_to_entries(entries, index, "vk_mv_percent_characteristic", vk_mv_percent_characteristic)
+    if vkr_mv_percent_characteristic is not None:
+        _add_series_to_entries(entries, index, "vkr_mv_percent_characteristic", vkr_mv_percent_characteristic)
+    if vk_lv_percent_characteristic is not None:
+        _add_series_to_entries(entries, index, "vk_lv_percent_characteristic", vk_lv_percent_characteristic)
+    if vkr_lv_percent_characteristic is not None:
+        _add_series_to_entries(entries, index, "vkr_lv_percent_characteristic", vkr_lv_percent_characteristic)
 
     _set_multiple_entries(net, "trafo3w", index, **entries, **kwargs)
 
@@ -3625,12 +3826,15 @@ def _create_column_and_set_value(net, index, variable, column, element, dtyp=flo
         set_value = True
     if set_value:
         if column not in net[element].columns:
-            if isinstance(default_val, str) \
-                    and version.parse(pd.__version__) < version.parse("1.0"):
-                net[element].loc[:, column] = pd.Series([default_val] * len(net[element]),
-                                                        dtype=dtyp)
-            else:
-                net[element].loc[:, column] = pd.Series(default_val, dtype=dtyp)
+            # # unclear why this if-clause is necessary, so commenting out for now
+            # if isinstance(default_val, str) \
+            #         and version.parse(pd.__version__) < version.parse("1.0"):
+            #     net[element].loc[:, column] = pd.Series([default_val] * len(net[element]),
+            #                                             dtype=dtyp)
+            # else:
+            # # this does not work properly (only sets 1 entry but not all to default_nan):
+            #     net[element].loc[:, column] = pd.Series(default_val, dtype=dtyp)
+            net[element].loc[:, column] = pd.Series(data=default_val, index=net[element].index, dtype=dtyp)
         net[element].at[index, column] = variable
     elif default_for_nan and column in net[element].columns:
         net[element].at[index, column] = default_val